Video-Trasnslation/pyproject.toml
[project]
name = "video-trasnslation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "googletrans>=4.0.2",
    "instructor>=1.7.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "youtube-transcript-api>=0.6.3",
]
